refactor: 이벤트 DTO를 common 패키지로 이동
