+++ src/main/java/kitchenpos/common/event/UpdateGroupOrderTableEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto;
+package kitchenpos.common.event;
 
 import java.util.List;
