+++ src/main/java/kitchenpos/common/event/ValidateSameSizeOrderTableEvent.java
@@ -1,12 +1,12 @@
-package kitchenpos.order.dto;
+package kitchenpos.common.event;
 
 import java.util.List;
 
-public class ValidateAppendOrderTableInTableGroupEvent {
+public class ValidateSameSizeOrderTableEvent {
 
     private final List<Long> orderTableIds;
 
-    public ValidateAppendOrderTableInTableGroupEvent(final List<Long> orderTableIds) {
+    public ValidateSameSizeOrderTableEvent(final List<Long> orderTableIds) {
         this.orderTableIds = orderTableIds;
     }
