+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -2,7 +2,7 @@ package kitchenpos.order.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.menu.dto.ValidateExistMenuEvent;
+import kitchenpos.common.event.ValidateExistMenuEvent;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
