+++ src/main/java/kitchenpos/order/application/TableEventHandler.java
@@ -5,11 +5,11 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.dto.UpdateGroupOrderTableEvent;
-import kitchenpos.order.dto.UpdateUngroupOrderTableEvent;
-import kitchenpos.order.dto.ValidateAppendOrderTableInTableGroupEvent;
-import kitchenpos.order.dto.ValidateOrderIsNotCompletionInOrderTableEvent;
-import kitchenpos.order.dto.ValidateSameSizeOrderTableEvent;
+import kitchenpos.common.event.UpdateGroupOrderTableEvent;
+import kitchenpos.common.event.UpdateUngroupOrderTableEvent;
+import kitchenpos.common.event.ValidateAppendOrderTableInTableGroupEvent;
+import kitchenpos.common.event.ValidateOrderIsNotCompletionInOrderTableEvent;
+import kitchenpos.common.event.ValidateSameSizeOrderTableEvent;
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.order.repository.OrderTableRepository;
 import org.springframework.context.event.EventListener;
