+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -3,11 +3,11 @@ package kitchenpos.tablegroup.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.order.dto.OrderTableInTableGroupDto;
-import kitchenpos.order.dto.UpdateGroupOrderTableEvent;
-import kitchenpos.order.dto.UpdateUngroupOrderTableEvent;
-import kitchenpos.order.dto.ValidateAppendOrderTableInTableGroupEvent;
-import kitchenpos.order.dto.ValidateOrderIsNotCompletionInOrderTableEvent;
-import kitchenpos.order.dto.ValidateSameSizeOrderTableEvent;
+import kitchenpos.common.event.UpdateGroupOrderTableEvent;
+import kitchenpos.common.event.UpdateUngroupOrderTableEvent;
+import kitchenpos.common.event.ValidateAppendOrderTableInTableGroupEvent;
+import kitchenpos.common.event.ValidateOrderIsNotCompletionInOrderTableEvent;
+import kitchenpos.common.event.ValidateSameSizeOrderTableEvent;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.tablegroup.dto.TableGroupCreateRequest;
 import kitchenpos.tablegroup.dto.TableGroupResponse;
